+++ Source/IGListKit/IGListSectionController.h
@@ -121,6 +121,40 @@ NS_SWIFT_NAME(ListSectionController)
 - (void)didUnhighlightItemAtIndex:(NSInteger)index;
 
 /**
+ Tells the section controller that the cell has requested a menu configuration.
+ 
+ @param index The index of the cell that requested the menu.
+ @param point The point of the tap on the cell.
+ 
+ @return An object that conforms to `UIContextMenuConfiguration`
+ 
+ @note The default implementation does nothing. **Calling super is not required.**
+ */
+- (UIContextMenuConfiguration * _Nullable)contextMenuConfigurationForItemAtIndex:(NSInteger)index point:(CGPoint)point API_AVAILABLE(ios(13.0)) API_UNAVAILABLE(tvos);
+
+/**
+ Tells the section controller that the cell has requested a preview for context menu highlight.
+ 
+ @param configuration Context menu configuration.
+ 
+ @return An object that conforms to `UITargetedPreview`
+ 
+ @note The default implementation does nothing. **Calling super is not required.**
+ */
+- (UITargetedPreview * _Nullable)previewForHighlightingContextMenuWithConfiguration:(UIContextMenuConfiguration *)configuration API_AVAILABLE(ios(13.0)) API_UNAVAILABLE(tvos);
+
+/**
+ Tells the section controller that the cell has requested a preview for context menu dismiss.
+ 
+ @param configuration Context menu configuration.
+ 
+ @return An object that conforms to `UITargetedPreview`
+ 
+ @note The default implementation does nothing. **Calling super is not required.**
+ */
+- (UITargetedPreview * _Nullable)previewForDismissingContextMenuWithConfiguration:(UIContextMenuConfiguration *)configuration API_AVAILABLE(ios(13.0)) API_UNAVAILABLE(tvos);
+
+/**
  Identifies whether an object can be moved through interactive reordering.
 
  @param index The index of the object in the list.