+++ Source/IGListKit/Internal/IGListAdapter+UICollectionView.m
@@ -18,6 +18,12 @@
 
 #import "IGListAdapterInternal.h"
 
+@interface IGListAdapter ()
+
+@property (nonatomic, strong) NSMapTable *configurationToSectionController;
+
+@end
+
 @implementation IGListAdapter (UICollectionView)
 
 #pragma mark - UICollectionViewDataSource
@@ -268,6 +274,47 @@
     IGListSectionController * sectionController = [self sectionControllerForSection:indexPath.section];
     [sectionController didUnhighlightItemAtIndex:indexPath.item];
 }
+
+#if !TARGET_OS_TV
+- (UIContextMenuConfiguration *)collectionView:(UICollectionView *)collectionView contextMenuConfigurationForItemAtIndexPath:(NSIndexPath *)indexPath point:(CGPoint)point API_AVAILABLE(ios(13.0)) {
+    // forward this method to the delegate b/c this implementation will steal the message from the proxy
+    id<UICollectionViewDelegate> collectionViewDelegate = self.collectionViewDelegate;
+    if ([collectionViewDelegate respondsToSelector:@selector(collectionView:contextMenuConfigurationForItemAtIndexPath:point:)]) {
+        return [collectionViewDelegate collectionView:collectionView contextMenuConfigurationForItemAtIndexPath:indexPath point:point];
+    }
+
+    IGListSectionController * sectionController = [self sectionControllerForSection:indexPath.section];
+    
+    UIContextMenuConfiguration *configuration = [sectionController contextMenuConfigurationForItemAtIndex:indexPath.item point:point];
+    if (configuration) {
+        if (!self.configurationToSectionController) {
+            self.configurationToSectionController = [NSMapTable weakToWeakObjectsMapTable];
+        }
+        [self.configurationToSectionController setObject:configuration forKey:sectionController];
+    }
+    return nil;
+}
+
+- (UITargetedPreview *)collectionView:(UICollectionView *)collectionView previewForHighlightingContextMenuWithConfiguration:(UIContextMenuConfiguration *)configuration  API_AVAILABLE(ios(13.0)){
+    id<UICollectionViewDelegate> collectionViewDelegate = self.collectionViewDelegate;
+    if ([collectionViewDelegate respondsToSelector:@selector(collectionView:previewForHighlightingContextMenuWithConfiguration:)]) {
+        return [collectionViewDelegate collectionView:collectionView previewForHighlightingContextMenuWithConfiguration:configuration];
+    }
+    
+    IGListSectionController * sectionController = [self.configurationToSectionController objectForKey:configuration];
+    return [sectionController previewForHighlightingContextMenuWithConfiguration:configuration];
+}
+
+- (UITargetedPreview *)collectionView:(UICollectionView *)collectionView previewForDismissingContextMenuWithConfiguration:(UIContextMenuConfiguration *)configuration  API_AVAILABLE(ios(13.0)){
+    id<UICollectionViewDelegate> collectionViewDelegate = self.collectionViewDelegate;
+    if ([collectionViewDelegate respondsToSelector:@selector(collectionView:previewForDismissingContextMenuWithConfiguration:)]) {
+        return [collectionViewDelegate collectionView:collectionView previewForDismissingContextMenuWithConfiguration:configuration];
+    }
+    
+    IGListSectionController * sectionController = [self.configurationToSectionController objectForKey:configuration];
+    return [sectionController previewForDismissingContextMenuWithConfiguration:configuration];
+}
+#endif
 
 #pragma mark - UICollectionViewDelegateFlowLayout