+++ Source/IGListKit/Internal/IGListAdapterProxy.m
@@ -33,6 +33,9 @@ static BOOL isInterceptedSelector(SEL sel) {
             sel == @selector(collectionView:didDeselectItemAtIndexPath:) ||
             sel == @selector(collectionView:didHighlightItemAtIndexPath:) ||
             sel == @selector(collectionView:didUnhighlightItemAtIndexPath:) ||
+            sel == @selector(collectionView:contextMenuConfigurationForItemAtIndexPath:point:) ||
+            sel == @selector(collectionView:previewForHighlightingContextMenuWithConfiguration:) ||
+            sel == @selector(collectionView:previewForDismissingContextMenuWithConfiguration:) ||
             // UICollectionViewDelegateFlowLayout
             sel == @selector(collectionView:layout:sizeForItemAtIndexPath:) ||
             sel == @selector(collectionView:layout:insetForSectionAtIndex:) ||